+++ features/broadcast/distsql/handler/src/main/java/org/apache/shardingsphere/broadcast/distsql/handler/converter/BroadcastRuleConfigurationToDistSQLConverter.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.broadcast.distsql.handler.provider;
+package org.apache.shardingsphere.broadcast.distsql.handler.converter;
 
 import com.google.common.base.Joiner;
 import org.apache.shardingsphere.broadcast.config.BroadcastRuleConfiguration;
@@ -26,14 +26,9 @@ import org.apache.shardingsphere.distsql.handler.engine.query.ral.convert.RuleCo
  */
 public final class BroadcastRuleConfigurationToDistSQLConverter implements RuleConfigurationToDistSQLConverter<BroadcastRuleConfiguration> {
     
-    public static final String CREATE_BROADCAST_TABLE_RULE = "CREATE BROADCAST TABLE RULE %s;";
-    
     @Override
     public String convert(final BroadcastRuleConfiguration ruleConfig) {
-        if (ruleConfig.getTables().isEmpty()) {
-            return "";
-        }
-        return String.format(CREATE_BROADCAST_TABLE_RULE, Joiner.on(",").join(ruleConfig.getTables()));
+        return ruleConfig.getTables().isEmpty() ? "" : String.format(BroadcastConvertDistSQLConstants.SQL_PATTERN, Joiner.on(",").join(ruleConfig.getTables()));
     }
     
     @Override