+++ features/readwrite-splitting/distsql/handler/src/main/java/org/apache/shardingsphere/readwritesplitting/distsql/handler/converter/ReadwriteSplittingRuleConfigurationToDistSQLConverter.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.readwritesplitting.distsql.handler.provider;
+package org.apache.shardingsphere.readwritesplitting.distsql.handler.converter;
 
 import com.google.common.base.Strings;
 import org.apache.shardingsphere.distsql.handler.constant.DistSQLConstants;
@@ -24,7 +24,6 @@ import org.apache.shardingsphere.distsql.handler.engine.query.ral.convert.RuleCo
 import org.apache.shardingsphere.infra.algorithm.core.config.AlgorithmConfiguration;
 import org.apache.shardingsphere.readwritesplitting.config.ReadwriteSplittingRuleConfiguration;
 import org.apache.shardingsphere.readwritesplitting.config.rule.ReadwriteSplittingDataSourceGroupRuleConfiguration;
-import org.apache.shardingsphere.readwritesplitting.distsql.handler.constant.ReadwriteSplittingDistSQLConstants;
 
 import java.util.Collection;
 import java.util.Iterator;
@@ -40,10 +39,10 @@ public final class ReadwriteSplittingRuleConfigurationToDistSQLConverter impleme
         if (ruleConfig.getDataSourceGroups().isEmpty()) {
             return "";
         }
-        StringBuilder result = new StringBuilder(ReadwriteSplittingDistSQLConstants.CREATE_READWRITE_SPLITTING_RULE);
+        StringBuilder result = new StringBuilder(ReadwriteSplittingConvertDistSQLConstants.CREATE_READWRITE_SPLITTING_RULE);
         Iterator<ReadwriteSplittingDataSourceGroupRuleConfiguration> iterator = ruleConfig.getDataSourceGroups().iterator();
         while (iterator.hasNext()) {
-            appendStaticReadWriteSplittingRule(ruleConfig.getLoadBalancers(), iterator.next(), result);
+            appendReadWriteSplittingRule(ruleConfig.getLoadBalancers(), iterator.next(), result);
             if (iterator.hasNext()) {
                 result.append(DistSQLConstants.COMMA);
             }
@@ -52,12 +51,12 @@ public final class ReadwriteSplittingRuleConfigurationToDistSQLConverter impleme
         return result.toString();
     }
     
-    private void appendStaticReadWriteSplittingRule(final Map<String, AlgorithmConfiguration> loadBalancers,
-                                                    final ReadwriteSplittingDataSourceGroupRuleConfiguration dataSourceGroupRuleConfig, final StringBuilder stringBuilder) {
+    private void appendReadWriteSplittingRule(final Map<String, AlgorithmConfiguration> loadBalancers,
+                                              final ReadwriteSplittingDataSourceGroupRuleConfiguration dataSourceGroupRuleConfig, final StringBuilder stringBuilder) {
         String readDataSourceNames = getReadDataSourceNames(dataSourceGroupRuleConfig.getReadDataSourceNames());
         String transactionalReadQueryStrategy = dataSourceGroupRuleConfig.getTransactionalReadQueryStrategy().name();
         String loadBalancerType = getLoadBalancerType(loadBalancers.get(dataSourceGroupRuleConfig.getLoadBalancerName()));
-        stringBuilder.append(String.format(ReadwriteSplittingDistSQLConstants.READWRITE_SPLITTING_FOR_STATIC,
+        stringBuilder.append(String.format(ReadwriteSplittingConvertDistSQLConstants.READWRITE_SPLITTING_RULE,
                 dataSourceGroupRuleConfig.getName(), dataSourceGroupRuleConfig.getWriteDataSourceName(), readDataSourceNames, transactionalReadQueryStrategy, loadBalancerType));
     }
     
@@ -65,7 +64,7 @@ public final class ReadwriteSplittingRuleConfigurationToDistSQLConverter impleme
         StringBuilder result = new StringBuilder();
         Iterator<String> iterator = readDataSourceNames.iterator();
         while (iterator.hasNext()) {
-            result.append(String.format(ReadwriteSplittingDistSQLConstants.READ_RESOURCE, iterator.next()));
+            result.append(String.format(ReadwriteSplittingConvertDistSQLConstants.READ_DATA_SOURCE, iterator.next()));
             if (iterator.hasNext()) {
                 result.append(DistSQLConstants.COMMA);
             }