+++ kernel/single/distsql/handler/src/main/java/org/apache/shardingsphere/single/distsql/handler/converter/SingleConvertDistSQLConstants.java
@@ -15,30 +15,18 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.rewrite.sql.token.keygen.pojo;
+package org.apache.shardingsphere.single.distsql.handler.converter;
 
-import org.apache.shardingsphere.infra.rewrite.sql.token.common.pojo.Attachable;
-import org.apache.shardingsphere.infra.rewrite.sql.token.common.pojo.SQLToken;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
 
 /**
- * Generated key insert column token.
+ * Single convert DistSQL constants.
  */
-public final class GeneratedKeyInsertColumnToken extends SQLToken implements Attachable {
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public final class SingleConvertDistSQLConstants {
     
-    private final String column;
+    public static final String LOAD_SQL_PATTERN = "LOAD SINGLE TABLE %s;";
     
-    public GeneratedKeyInsertColumnToken(final int startIndex, final String column) {
-        super(startIndex);
-        this.column = column;
-    }
-    
-    @Override
-    public String toString() {
-        return ", " + column;
-    }
-    
-    @Override
-    public int getStopIndex() {
-        return getStartIndex();
-    }
+    public static final String SET_DEFAULT_SQL_PATTERN = "SET DEFAULT SINGLE TABLE STORAGE UNIT = %s;";
 }