+++ kernel/single/distsql/handler/src/main/java/org/apache/shardingsphere/single/distsql/handler/converter/SingleRuleConfigurationToDistSQLConverter.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.single.distsql.handler.provider;
+package org.apache.shardingsphere.single.distsql.handler.converter;
 
 import com.google.common.base.Joiner;
 import org.apache.shardingsphere.distsql.handler.engine.query.ral.convert.RuleConfigurationToDistSQLConverter;
@@ -26,15 +26,8 @@ import org.apache.shardingsphere.single.config.SingleRuleConfiguration;
  */
 public final class SingleRuleConfigurationToDistSQLConverter implements RuleConfigurationToDistSQLConverter<SingleRuleConfiguration> {
     
-    public static final String LOAD_SINGLE_TABLE = "LOAD SINGLE TABLE %s;";
-    
-    public static final String SET_DEFAULT_SINGLE_TABLE_STORAGE_UNIT = "SET DEFAULT SINGLE TABLE STORAGE UNIT = %s;";
-    
     @Override
     public String convert(final SingleRuleConfiguration ruleConfig) {
-        if (ruleConfig.getTables().isEmpty() && !ruleConfig.getDefaultDataSource().isPresent()) {
-            return "";
-        }
         StringBuilder result = new StringBuilder();
         if (!ruleConfig.getTables().isEmpty()) {
             result.append(convertLoadTable(ruleConfig));
@@ -49,11 +42,11 @@ public final class SingleRuleConfigurationToDistSQLConverter implements RuleConf
     }
     
     private String convertLoadTable(final SingleRuleConfiguration ruleConfig) {
-        return String.format(LOAD_SINGLE_TABLE, Joiner.on(",").join(ruleConfig.getTables()));
+        return String.format(SingleConvertDistSQLConstants.LOAD_SQL_PATTERN, Joiner.on(",").join(ruleConfig.getTables()));
     }
     
     private String convertSetDefaultSingleTableStorageUnit(final String defaultStorageUnitName) {
-        return String.format(SET_DEFAULT_SINGLE_TABLE_STORAGE_UNIT, defaultStorageUnitName);
+        return String.format(SingleConvertDistSQLConstants.SET_DEFAULT_SQL_PATTERN, defaultStorageUnitName);
     }
     
     @Override